several updates on movie form view

+++ AbcMovies/Models/Movie.cs
@@ -24,7 +24,7 @@ namespace AbcMovies.Models
 
         [Required]
         [Display(Name = "Genre")]
-        public int GenreId { get; set; }
+        public byte GenreId { get; set; }
 
 
     }
